Adds trip controller
